Split main into main_start_up, main_clean_up, and main_loop.
The way it is now structured, we will continue on even if there is an error
cleaning up. The final result will then be a collect from the iterator of
Results, so if the main loop had an error or any of the clean up functions had
an error, we will return an error.