Implement containers widget myself.
Before, we were just using the List widget with a ListState variable. But, we
want to have more control over the widget: scrolling, where the cursor is, etc.
To do that, we move to a model where we track the viewport and selection
ourself, and then just provide a stateless List widget with the items that fit
in the viewport.