Add a timer task.
Messages will be sent to the timer task to start a timer or cancel one. The
timer task will send messages when timers expire.

The implementation of the timer is very basic as we don't expect there to be
many timers outstanding at any given time. A more industrial-strength
implementation would use a heap.